questao01/requirements.txt
requests==2.28.2
PyNaCl==1.5.0
Scrapy==2.8.0
python-dotenv==1.0.0